pyproject.toml
[tool.poetry]
name = "HW14_Deploy"
version = "0.1.0"
description = ""
authors = ["AnybodyKnows <120731166+AnybodyKnows@users.noreply.github.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
unicorn = {extras = ["standard"], version = "^2.0.1.post1"}
sqlalchemy = {extras = ["postgresql-asyncpg"], version = "^2.0.31"}
alembic = "^1.13.2"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pathlib = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
libgravatar = "^1.0.4"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "==4.*"
fastapi-limiter = "^0.1.6"
jinja2 = "^3.1.4"
cloudinary = "^1.41.0"
uvicorn = "^0.30.5"
pytest = "^8.3.2"


[tool.poetry.group.test.dependencies]
aiosqlite = "^0.20.0"
pytest-asyncio = "^0.23.8"
httpx = "^0.27.0"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["tests",]
pythonpath = "."
filterwarnings = "ignore::DeprecationWarning"